+++ include/bitsery/common.h
@@ -27,6 +27,8 @@
 #include "details/buffer_common.h"
 #include <vector>
 
+#include <list>
+
 namespace bitsery {
 
     struct DefaultConfig {